+++ exampleWithEmbedding.py
@@ -6,6 +6,8 @@
 import json
 import time
 
+from tcvectordb import exceptions
+
 import tcvectordb
 from tcvectordb.model.collection import Embedding
 from tcvectordb.model.document import Document, Filter, SearchParams, HNSWSearchParams
@@ -35,8 +37,12 @@ class TestVDB:
                                                  read_consistency=ReadConsistency.EVENTUAL_CONSISTENCY, timeout=timeout)
 
     def clear(self):
-        db = self._client.database('book')
-        db.drop_database('book')
+        try:
+            db = self._client.database('book')
+            if db:
+                db.drop_database('book')
+        except exceptions.ParamError:
+            pass
 
     def delete_and_drop(self):
         db = self._client.database('book')
